master Refactor tests structure
- add files for inner classses
