+++ retrofireTests/RemoteBaseTest.swift
@@ -175,115 +175,3 @@ class RemoteBaseTest: QuickSpec {
         
     }
 }
-
-
-/**
- Class implementation of RemoteBase to use for tests
- */
-class RemoteBaseImpl: RemoteBase {
-    
-    func constructionSites() -> Call<[ResponseObject]> {
-        let path = "http://stage.stant.com.br/api/v1/construction_sites"
-        let request = RequestBuilder(path: path)
-            .headers(["Authorization": "Beare some-beaerer"])
-            .build()
-        return self
-            .withCustomErrorClass(CustomErrorResponse.self)
-            .callList(request: request)
-    }
-    
-    func posts() -> Call<[ResponseObject]> {
-        let path = "https://jsonplaceholder.typicode.com/posts"
-        let request = RequestBuilder(path: path).build()
-        return self.callList(request: request)
-    }
-    
-    func postComments(postId: Int) -> Call<[ResponseObject]> {
-        let path = "http://jsonplaceholder.typicode.com/comments"
-        let request = RequestBuilder(path: path)
-            .queryParameters(["postId": postId.description])
-            .build()
-        return self.callList(request: request)
-    }
-    
-    func postsComments(postId: Int, email: String) -> Call<[ResponseObject]> {
-        let path = "http://jsonplaceholder.typicode.com/comments"
-        let request = RequestBuilder(path: path)
-            .queryParameters(["postId": postId.description, "email": email])
-            .build()
-        return self.callList(request: request)
-    }
-    
-    func findPost(id: Int) -> Call<ResponseObject> {
-        let path = "http://jsonplaceholder.typicode.com/posts/\(id)/"
-        let request = RequestBuilder(path: path)
-            .build()
-        return self.callSingle(request: request)
-    }
-    
-    func createPost(userId: Int, title: String, body: String) -> Call<ResponseObject> {
-        let path = "http://jsonplaceholder.typicode.com/posts"
-        let request = RequestBuilder(path: path)
-            .method(.post)
-            .bodyParameters(["userId": userId, "title": title, "body": body])
-            .build()
-        return self.callSingle(request: request)
-    }
-    
-    func updatePost(id: Int, userId: Int, title: String, body: String) -> Call<ResponseObject> {
-        let path = "http://jsonplaceholder.typicode.com/posts/\(id)"
-        let request = RequestBuilder(path: path)
-            .method(.put)
-            .bodyParameters(["userId": userId, "title": title, "body": body])
-            .build()
-        return self.callSingle(request: request)
-    }
-    
-    func deletePost(id: Int) -> Call<Default> {
-        let path = "http://jsonplaceholder.typicode.com/posts/\(id)"
-        let request = RequestBuilder(path: path)
-            .method(.delete)
-            .build()
-        return self.callSingle(request: request)
-    }
-    
-}
-
-/**
- Struct to use as Response on tests
- */
-import SwiftyJSON
-
-struct ResponseObject: Retrofire.Mappable {
-    let userId: Int?
-    let id: Int?
-    let title: String?
-    let body: String?
-    
-    static func instanceBy<M>(json: JSON) -> M {
-        let userId = json.dictionary?["userId"]?.int
-        let id     = json.dictionary?["id"]?.int
-        let title  = json.dictionary?["title"]?.string
-        let body   = json.dictionary?["body"]?.string
-        
-        return ResponseObject.init(userId: userId, id: id, title: title, body: body) as! M
-    }
-}
-
-/**
- Struct to use as ErrorResponse on tests
- */
-
-struct CustomErrorResponse: Retrofire.Mappable {
-    let responseType: String?
-    let message: String?
-    let trace: [String]?
-    
-    static func instanceBy<M>(json: JSON) -> M {
-        let responseType = json.dictionary?["response_type"]?.stringValue
-        let message      = json.dictionary?["message"]?.stringValue
-        let trace        = json.dictionary?["trace"]?.arrayObject as? [String]
-        
-        return CustomErrorResponse.init(responseType: responseType, message: message, trace: trace) as! M
-    }
-}
